R339 回溯最长三年加半年档 + 卖出上色
用户: 「回溯时间太长了, 只看近三年和短周期, 你调整周期, 卖出也要上色,
这样看起来醒目」。

一、周期
[1,2,3,5] → [0.5,1,2,3], 不足一年按月说。后端 years 从 int/ge=1 放到
float/ge=0.5 —— 它只被 _load_batch 拿去算取多少天日线, 小数一路算得通,
放开的是取数窗口, 不是任何判定口径。

这一条推翻了我自己在 R332 立的守卫(「回溯给的是样本量, 不该为了看当下
砍短」)。论点没错, 但它替用户做了决定。守卫改成钉新口径, 而样本量那件事
改由界面自己说: 完整买卖少于 10 轮时胜率标黄, 写明「只有 N 轮, 样本太少
不当数」。半年约 120 个交易日, 转折是低频信号, 那种窗口里可能只有两三次
完整买卖 —— 不是不好看, 是不成立; 与其让测试拦着用户, 不如把话摆在数字
旁边。

二、上色
在这之前买和卖共用同一个灰蓝底 bg-accent/[0.06]。徽标虽然分了红绿, 但
一行里最先被看见的是整条底色, 而底色对买和卖说的是同一句话。现在底色
跟着方向走(买 bull / 卖 bear), 各加一道 2px 左色条。

另加一档「贴着清仓线」: 手上的票离离场线 2% 以内同样上琥珀 —— 卖出这一侧
该醒目的不只是「今天要卖」, 还有「明后天很可能要卖」。这一档不带动作徽标,
判据写成 !actionable && r.held && …, 它一旦能脱离 !actionable 独立成立
就是偷偷多了一个出手口径, 守卫直接钉这一行。

14 处变异逐一确认会红。

写守卫时翻了一次车并留在注释里: 两个模块都开了 from __future__ import
annotations, 注解全是字符串, inspect.signature(...).annotation is float
永远为假, 要 get_type_hints 才算数。

后端 pytest 3937 passed / 10 skipped 退出码 0;
前端 tsc 0 / eslint 0 error / build 通过; 作者层与打分模块 diff 全空。

+++ backend/app/services/flip_portfolio_run.py
@@ -40,7 +40,7 @@ _COLS = ("symbol", "date", "close", "signal_limit_up", "signal_limit_down")
 def run(repo, *, symbols: list[str] | None = None,
         capital: float = flip_portfolio.DEFAULT_CAPITAL,
         max_positions: int = flip_portfolio.DEFAULT_MAX_POSITIONS,
-        years: int = DEFAULT_YEARS) -> dict:
+        years: float = DEFAULT_YEARS) -> dict:
     """跑一遍转折模拟盘。`symbols` 不给就取当前自选。"""
     # `symbols is None` = 没指定, 用自选; `symbols == []` = **明确要空**, 就是空。
     # 用 `symbols or 自选` 会把这两件事混成一件 —— 「没给」与「给了个空的」不是
@@ -114,7 +114,7 @@ def _names(repo, syms: list[str]) -> dict[str, str]:
     return {s: str(m.get(s) or s) for s in syms}
 
 
-def _load_batch(repo, syms: list[str], years: int) -> dict[str, pl.DataFrame]:
+def _load_batch(repo, syms: list[str], years: float) -> dict[str, pl.DataFrame]:
     """一趟批量读。非股票(ETF/指数)批量接口给不了, 逐只回退 —— **不静默丢掉**。"""
     end = date.today()
     span = int((years * 250 + _WARMUP_BARS) * _CALENDAR_RATIO) + 30
